Context menu refactor (provisional draft branch) (#5225)
* context_menu_gpx_appearean_tags_fix

* rewrite OAGpxAppearanceInfo to swift

* remove old files

* quickfix

* Gpx appearance info sync (#5031)

* context_menu_gpx_appearean_tags_fix

* clean code

* manual add file

* rewrite OAGpxAppearanceInfo to swift

* remove old files

* quickfix

* sync GpxAppearanceInfo with android

* sync with android

* code review fix

* Fix Bool value

* Add private static func boolValue

---------


* fix conflict

* fix conflict

* fix conflict

* fix conflict

* merge conflicts fixed. code cleared

* code cleared

* clear code

* code update

* crash fix. code update

* crash fix

* Synchronize BaseDetailsObject

* Remove whitespaces

* review fix

* remove duplicated variable

* sync with android. it can contain uninited latlon

* remove unused code

* review fix

* review fix

* review fix

* review fix

* Add TODO for possible double call

* review fix

* review fix

* review fix

* Add TODO @RZR

* code style fix

* Add OATargetBaseDetailsObject OATargetRenderedObject

* Add TODO

* React on OATargetRenderedObject in OATargetMenuViewController

* Add searchDetailedObject to RenderedObjectViewController

* Fix coordinayes

* Fix comparison with CLLocationCoordinate2DIsValid

* Fix combine POI from search

* Sync MapSelectionHelper allowMapObjects with Java

* Fix AmenitySearcher

* some AmenityUIHelper syncing. refactor

* sync some poi row style helpers

* context menu implementation

* context menu implementation

* context menu implementation

* MergeLocalizedTagsAlgorithm bugfix

* icons fix

* fix formatting

* some collapsed view implementation

* implement some multiple values cell

* clean code

* clean code

* code clear. refactor

* add names row. refactor

* add travel guides row

* Fix detailed object in RenderedObjectViewController

* wikidata row fix opening url

* wikidata row fix opening url

* deleted old unused code

* Fix in open Names Row

* wiki description added

* clean code. fix some todo

* compilation fix initWithMapObjects

* Fix

* refactoring

* refactoring

* add - (int)pointOrder:(id)object

* Fix duplicates

* add createMapObjectIdFromCleanOsmId

* Fix POIUiFilter

* fix startLoadingImages

* Order of obfs

* Fix

* fix merge conflicts

* fix obfID and osmID null checkings

* todo fix

* todo fix. refactor

* typo fix

* travel guides row fix

* imolement PlaceDetailsViewController. buildDescription()

* refactor

* typo fix

* fix contextMenuDidShow

* clean code

* fixes

* Fix popular places

* mapPanel.updateContextMenu

* updateTargetPoint

* fix: use detailsObject for targetMenuView (title/icon)

* replacingNbsp

* avoid building cells twice in buildMenu for FavoriteViewController

* use searchDetailedObject for OATargetBaseDetailsObject | OATargetRenderedObject

* fix transportStop visualization

* Fix wikiwoyage merging

* grooming

* crash fix

* clear and grooming

* add row createWikipediaArticleList

* fix parsing poiCategory for AmenityUIHelper

* Fix crash on OSM notes row tap

* fix tap OSM nots on map

* grooming

* pointOrder -> pointsOrder and OAOpenedHoursParser -> OAOpeningHoursParser

---------

+++ Sources/OsmEdit/OAOpenStreetMapLocalUtil.mm
@@ -65,7 +65,7 @@
 - (OAEntity *)loadEntity:(OATargetPoint *)targetPoint
 {
     OAPOIHelper *poiHelper = [OAPOIHelper sharedInstance];
-    long long objectId = targetPoint.obfId;
+    uint64_t objectId = targetPoint.obfId;
     BOOL isTransportStop = targetPoint.type == OATargetTransportStop;
     if (isTransportStop)
         objectId = ((OATransportStop *)targetPoint.targetObj).poi.obfId;
@@ -78,7 +78,7 @@
     OAPOIType *poiType = poi.type;
     BOOL isAmenity = poiType && ![poiType isKindOfClass:[OAPOILocationType class]];
     
-    long long entityId = objectId >> (isAmenity ? AMENITY_ID_RIGHT_SHIFT : NON_AMENITY_ID_RIGHT_SHIFT);
+    uint64_t entityId = objectId >> (isAmenity ? AMENITY_ID_RIGHT_SHIFT : NON_AMENITY_ID_RIGHT_SHIFT);
     BOOL isWay = objectId % 2 == WAY_MODULO_REMAINDER; // check if mapObject is a way
     
     OAEntity *entity;
